requirements.txt
# === Core Backend ===
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
requests==2.32.3
aiohttp==3.10.3
rich==13.9.2

# === Google Gemini / LLM ===
google-generativeai==0.8.3

# === Machine Learning & Data Science ===
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
scipy==1.13.1
xgboost==2.1.1
gensim==4.3.3
joblib==1.4.2
matplotlib==3.9.2
seaborn==0.13.2

# === Logging & Utilities ===
tqdm==4.66.5
pathlib==1.0.1

# === Testing / Development ===
pytest==8.3.3
httpx==0.27.2
black==24.10.0
isort==5.13.2
flake8==7.1.1
